# TRAAN_Integration_Phase5/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SIH Disaster Response API",
    "description": "Disaster Response Coordination Platform \u2014 Backend API\n\n**SIH 2026** \u00b7 Component D \u2014 Backend Core\n\nThis Swagger UI is the live contract that Android, Dashboard, and AI/Optimizer teams build against. Do not change endpoint paths, request fields, or response fields without coordinating with all affected teams.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/resources": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Get Resources",
        "description": "List resources with optional filtering.",
        "operationId": "get_resources_api_v1_resources_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "district",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResourceResponse"
                  },
                  "title": "Response Get Resources Api V1 Resources Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "resources"
        ],
        "summary": "Create Resource",
        "description": "Create a new mock IDRN resource. Admin only.",
        "operationId": "create_resource_api_v1_resources_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/{resource_id}": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Get Resource",
        "operationId": "get_resource_api_v1_resources__resource_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "resources"
        ],
        "summary": "Update Resource",
        "description": "Update a resource (e.g. quantity or status). Authority only.",
        "operationId": "update_resource_api_v1_resources__resource_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/situation-brief": {
      "get": {
        "tags": [
          "optimizer"
        ],
        "summary": "Get Situation Brief",
        "description": "STUB: Returns the current situation brief.\nComponent E is responsible for the actual Gemini implementation.",
        "operationId": "get_situation_brief_api_v1_situation_brief_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/situation-brief/refresh": {
      "post": {
        "tags": [
          "optimizer"
        ],
        "summary": "Refresh Situation Brief",
        "description": "STUB: Refreshes the situation brief.\nComponent E is responsible for the actual Gemini implementation.",
        "operationId": "refresh_situation_brief_api_v1_situation_brief_refresh_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/optimize/allocate": {
      "post": {
        "tags": [
          "optimizer"
        ],
        "summary": "Allocate Resources",
        "description": "STUB: Requests resource allocation.\nComponent E is responsible for the actual OR-Tools implementation.",
        "operationId": "allocate_resources_api_v1_optimize_allocate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllocateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "AllocateRequest": {
        "properties": {
          "incident_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Ids"
          },
          "constraints": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constraints"
          }
        },
        "type": "object",
        "title": "AllocateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResourceCreate": {
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "medical",
              "rescue",
              "shelter",
              "transport",
              "communication"
            ],
            "title": "Category"
          },
          "sub_type": {
            "type": "string",
            "title": "Sub Type"
          },
          "custodian_agency": {
            "type": "string",
            "title": "Custodian Agency"
          },
          "quantity_total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Quantity Total"
          },
          "quantity_available": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Quantity Available"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "partially_deployed",
              "deployed",
              "maintenance"
            ],
            "title": "Status"
          },
          "contact": {
            "type": "string",
            "title": "Contact"
          },
          "location": {
            "$ref": "#/components/schemas/app__models__resources__Location"
          }
        },
        "type": "object",
        "required": [
          "category",
          "sub_type",
          "custodian_agency",
          "quantity_total",
          "quantity_available",
          "status",
          "contact",
          "location"
        ],
        "title": "ResourceCreate"
      },
      "ResourceResponse": {
        "properties": {
          "category": {
            "type": "string",
            "enum": [
              "medical",
              "rescue",
              "shelter",
              "transport",
              "communication"
            ],
            "title": "Category"
          },
          "sub_type": {
            "type": "string",
            "title": "Sub Type"
          },
          "custodian_agency": {
            "type": "string",
            "title": "Custodian Agency"
          },
          "quantity_total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Quantity Total"
          },
          "quantity_available": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Quantity Available"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "partially_deployed",
              "deployed",
              "maintenance"
            ],
            "title": "Status"
          },
          "contact": {
            "type": "string",
            "title": "Contact"
          },
          "resource_id": {
            "type": "string",
            "title": "Resource Id"
          },
          "location": {
            "$ref": "#/components/schemas/app__models__resources__Location"
          },
          "last_updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Updated At"
          }
        },
        "type": "object",
        "required": [
          "category",
          "sub_type",
          "custodian_agency",
          "quantity_total",
          "quantity_available",
          "status",
          "contact",
          "resource_id",
          "location",
          "last_updated_at"
        ],
        "title": "ResourceResponse"
      },
...
      "ResourceUpdate": {
        "properties": {
          "quantity_available": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity Available"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "ResourceUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "app__models__resources__Location": {
        "properties": {
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng"
          },
          "district": {
            "type": "string",
            "title": "District"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lng",
          "district"
        ],
        "title": "Location"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}